Converted DocumentToAssetTest from Integration to Unit Test
